refactor: use discord.js 14 functionality to set, clear, fetch commands and command permissions
