+++ src/core/permissions/AdminPermissions.ts
@@ -1,7 +1,7 @@
 import * as Discord from "discord.js";
 
 import Logger from "../../log";
-import DiscordPermissionsV2Utils from "../../util/discord-patch/DiscordPermissionsV2Utils";
+import { canUseCommand } from "../../util/permissions";
 import InteractionRegistry from "../InteractionRegistry";
 import { CustomSample } from "../soundboard/CustomSample";
 import { StandardSample } from "../soundboard/StandardSample";
@@ -11,22 +11,20 @@ const log = Logger.child({ label: "AdminPermissions" });
 export default class AdminPermissions {
     public client: Discord.Client<true>;
     public registry: InteractionRegistry;
-    public discord: DiscordPermissionsV2Utils;
 
     constructor(client: Discord.Client<true>, registry: InteractionRegistry) {
         this.client = client;
         this.registry = registry;
-        this.discord = DiscordPermissionsV2Utils.client(client);
     }
 
     public async isAdmin(guild: Discord.Guild, userId: Discord.Snowflake): Promise<boolean> {
-        const config_api_command = this.registry.getAPICommand(guild.id, "config");
-        if (!config_api_command) {
+        const config_command = this.registry.getApplicationCommand(guild.id, "config");
+        if (!config_command) {
             log.error("Command couldn't be found in registry", { guildId: guild.id, commandName: "config" });
             return false;
         }
 
-        const admin_permissions = await this.discord.canUseCommand(config_api_command, guild, null, userId);
+        const admin_permissions = await canUseCommand(config_command, guild, null, userId);
 
         return admin_permissions.user;
     }
